Refactor common code into images.py functions
