+++ src/js/components/service-requests/service-request-detail.js
@@ -8,9 +8,7 @@ import { BlockPreWrap } from "../atoms/block-pre-wrap";
 import { FullWidthGrid } from "../grid";
 import { LoadingPlaceholder } from "../atoms/loading-placeholder";
 import { StatusMessage } from "../molecules/status-message";
-import { v4 as uuidv4 } from "uuid";
-import { getLabel } from "../../utils/element-factory";
-import { ServiceRequestSubmitted } from "./service-request-submitted";
+import { createImageFormFields, updateUploadedFiles } from "./images";
 
 export class ServiceRequestDetail {
   constructor() {
@@ -22,25 +20,12 @@ export class ServiceRequestDetail {
     const serviceRequestId = new URLSearchParams(url.search).get("id");
     const $sectionHeading = $(".components .section-heading");
 
-    const $formInputTmpl = $(".components .form__input-field:eq(0)");
-
-    const $uploadImagesLabel = getLabel("Images of your issue");
-    const $uploadImagesInput = $formInputTmpl.clone().attr({
-      id: "upload-images-input",
-      name: "files",
-      type: "file",
-      accept: "image/*",
-      multiple: true,
-      style: "display: none",
-    });
-    const $uploadImagesClass = $(".upload-images");
-    const $uploadImagePreview = $(".image-preview");
-    const $uploadImageAdd = $(".button.button--add-image");
-
-    // Reroute the click event from the Add element to the html input element
-    $uploadImageAdd.click(function () {
-      $uploadImagesInput.click();
-    });
+    const {
+      $uploadImagesInput,
+      $uploadImagesLabel,
+      $uploadImagesClass,
+      $uploadImagePreview,
+    } = createImageFormFields();
 
     /* We keep a mapping of uploaded images separate from the FileList stored on
     the input object because we want previously selected images to persist after
@@ -55,49 +40,15 @@ export class ServiceRequestDetail {
      */
     let uploadedFiles = {};
 
-    function handleFileUpload() {
-      for (let i = 0; i < this.files.length; i++) {
-        let uuid = uuidv4();
-        const $preview = $uploadImagePreview
-          .clone()
-          .attr({
-            id: "upload-image-preview-" + uuid,
-          })
-          .removeClass("hidden");
-
-        const $previewRemove = $preview.find(".image-preview__remove");
-        $previewRemove.click(function () {
-          delete uploadedFiles[uuid];
-          $("#upload-image-preview-" + uuid).remove();
-
-          if (Object.keys(uploadedFiles).length > 0) {
-            $("#submit-images").show();
-          } else {
-            $("#submit-images").hide();
-          }
-        });
-
-        let reader = new FileReader();
-        reader.onload = function (e) {
-          // Replace newlines in base64 encoding so it doesn't break CSS
-          $preview.css(
-            "background-image",
-            "url('" + e.target.result.replace(/(\r\n|\n|\r)/gm, "") + "')"
-          );
-        };
-        reader.readAsDataURL(this.files[i]);
-        uploadedFiles[uuid] = this.files[i];
-        $uploadImagesClass.append($preview);
-
-        if (Object.keys(uploadedFiles).length > 0) {
-          $("#submit-images").show();
-        } else {
-          $("#submit-images").hide();
-        }
-      }
+    function handleInputFilesChanged() {
+      updateUploadedFiles(
+        this.files,
+        uploadedFiles,
+        $uploadImagePreview,
+        $uploadImagesClass
+      );
     }
-
-    $uploadImagesInput.change(handleFileUpload);
+    $uploadImagesInput.change(handleInputFilesChanged);
 
     const $submitButton = $(".components .button.button--form-submit")
       .clone()
@@ -112,6 +63,7 @@ export class ServiceRequestDetail {
       event.preventDefault();
       if (Object.keys(uploadedFiles).length === 0) {
         alert("Please select one or more images to upload.");
+        return;
       }
 
       let formData = new FormData();
@@ -127,7 +79,7 @@ export class ServiceRequestDetail {
         .fail((a, b) => {
           this.$element.empty().append(
             new StatusMessage({
-              text: "Error while updating service request.",
+              text: "Error while adding files to the service request.",
               status: "failure",
             }).render()
           );
